requirements.txt
click==8.1.7
requests==2.31.0
PyYAML==6.0.1
paramiko==3.4.0
rich==13.7.0